Check slob

Checks whether an object is a slob (serialized blob).

Usage

chk_slob(x, x_name = NULL)

Arguments

x

The object to check.

x_name

A string of the name of object x or NULL.

Value

NULL, invisibly. Called for the side effect of throwing an error if the condition is not met.

See also

Examples

chk_slob(flobr:::slob_obj)