flobr: Convert Files to and from Binary Objects (BLOBs)

Converts files to and from flobs. A flob is a file that was read into binary in integer-mode as little endian, saved as the single element of a named list (where the name is the name of the original file) and then serialized before being coerced into a blob. Flobs are useful for writing and reading files to and from databases.

See also

Author

Maintainer: Evan Amies-Galonski evan@poissonconsulting.ca (ORCID)

Authors:

Joe Thorley joe@poissonconsulting.ca (ORCID)

Other contributors:

Poisson Consulting [copyright holder, funder]